pyproject.toml
[tool.poetry]
name = "customer_api"
version = "0.1.0"
description = ""
authors = ["Thilakaraj K <nmrkthilakraj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.2"
Django = "^3.1.6"
psycopg2 = "^2.8.6"
djangorestframework-jwt = "^1.11.0"
requests = "^2.25.1"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
